test: add coverage for headless.ts
