+++ apps/storefront/tests/setup-test-environment.ts
@@ -6,8 +6,8 @@ import '@testing-library/jest-dom/vitest';
 
 window.B3 = {
   setting: {
-    channel_id: Number.NaN,
-    store_hash: '',
+    channel_id: 1,
+    store_hash: 'store-hash',
     platform: 'bigcommerce',
     environment: Environment.Local,
   },
@@ -15,8 +15,8 @@ window.B3 = {
 beforeEach(() => {
   window.B3 = {
     setting: {
-      channel_id: Number.NaN,
-      store_hash: '',
+      channel_id: 1,
+      store_hash: 'store-hash',
       platform: 'bigcommerce',
       environment: Environment.Local,
     },
